+++ lib/keyboard_custom.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
 /// Signature for a function that creates a widget for a given value
-typedef WidgetBuilder<T> = Widget Function(BuildContext context, T value);
+typedef WidgetBuilder<T> = Widget Function(
+    BuildContext context, T value, bool hasFocus);
 
 /// A widget that allow us to create a custom keyboard instead of the platform keyboard.
 class KeyboardCustomInput<T> extends StatefulWidget {
@@ -60,8 +61,8 @@ class _KeyboardCustomInputState<T> extends State<KeyboardCustomInput<T>> {
                 child: Container(
                   child: AnimatedBuilder(
                     animation: widget.notifier,
-                    builder: (context, child) =>
-                        widget.builder(context, widget.notifier.value),
+                    builder: (context, child) => widget.builder(
+                        context, widget.notifier.value, _hasFocus),
                   ),
                 ),
               ),